packages/ingestion/requirements.txt
psycopg2-binary>=2.9
pytest>=7.0
python-dotenv
